+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Bug
 - \#168 - fix incorrect demo in readme
+- \#172 - allow empty list in flush passing parameter
+
+## New Feature
+- \#170 - add index RNSG into HTTP handler
+- \#171 - add two new binary metric: SUBSTRUCTURE and SUPERSTRUCTURE
 
 
 # pymilvus 0.2.8(2020-03-11)